feat: refactor: 바뀐 endpoint에 맞게 user api 수정(onboarding repo 삭제)
