+++ src/api/index.ts
@@ -2,6 +2,5 @@ export { default as API } from './API';
 export { default as AuthRepository } from './AuthRepository';
 export { default as CommentRepository } from './CommentRepository';
 export { default as FeedRepository } from './FeedRepository';
-export { default as OnboardingRepository } from './OnboardingRepository';
 export { default as UserRepository } from './UserRepository'
 export * from './queryKeys';
