+++ src/types/index.ts
@@ -1,5 +1,5 @@
 export * as Auth from './AuthDTO';
 export * as Comment from './CommentDTO';
 export * as Feed from './FeedDTO';
-export * as Onboarding from './OnboardingDTO';
+export * as User from './UserDTO';
 export * from './ImageFile';
